Make GenerateSourcesTask easier to extend.
