+++ src/main/java/net/fabricmc/loom/configuration/decompile/SingleJarDecompileConfiguration.java
@@ -32,7 +32,6 @@ import org.gradle.api.Project;
 
 import net.fabricmc.loom.LoomGradleExtension;
 import net.fabricmc.loom.configuration.providers.minecraft.mapped.MappedMinecraftProvider;
-import net.fabricmc.loom.task.GenerateSourcesTask;
 import net.fabricmc.loom.util.Constants;
 
 public class SingleJarDecompileConfiguration extends DecompileConfiguration<MappedMinecraftProvider> {
@@ -62,7 +61,7 @@ public class SingleJarDecompileConfiguration extends DecompileConfiguration<Mapp
 			final String decompilerName = options.getName().substring(0, 1).toUpperCase() + options.getName().substring(1);
 			String taskName = "genSourcesWith" + decompilerName;
 			// Decompiler will be passed to the constructor of GenerateSourcesTask
-			project.getTasks().register(taskName, GenerateSourcesTask.class, options).configure(task -> {
+			project.getTasks().register(taskName, options.getTaskClass().get(), options).configure(task -> {
 				task.getInputJar().set(inputJar);
 				task.getRuntimeJar().set(namedJar);
